requirements.txt
Django==2.2.24
numpy==1.21.3
pandas==1.3.4
python-dateutil==2.8.2
mysqlclient==2.0.3
django-mysql==3.12.0
djangorestframework==3.9.4
pandas==1.3.4